Refactored assertThingProperty methods to reduce code.
The eight methods spanning the four Model Domain contexts are all
performing very similar tasks, and will only continue to grow in
number as the Model is expanded.

To reduce the amount of code needed to be maintained, and to create
a more general purpose re-usable method, I implemented a method
that allows the developer to assert the value of any property of any
object in the Store.