Refactored ensureThingExists methods to reduce code.
The four methods spanning the four Model Domain contexts are all
performing very similar tasks, and will only continue to grow in
number as the Model is expanded.

To reduce the amount of code needed to be maintained, and to create
a more general purpose re-usable method, I implemented a method
that allows the developer to ensure that an instance of any class
with a zero argument constructor exists.